fix(stablehlo): address copilot quantization feedback
